Move server, take notes
