+++ clueless/static/gameroom.html
@@ -87,9 +87,6 @@
       </div>
     </form>
 
-    <h3>Message Complete</h3>
-    <p>TBD</p>
-
     <h3>Message Move</h3>
     <form class="form" id="formMove">
       <p>Choose a position between 0 and 20.</p>
@@ -149,7 +146,7 @@
     <hr>
 
     <h1>Skeletal Increment: Messaging Demonstration</h1>
-    <p>In the below, each message defined in our SRS can be exercised usinga button (and input fields when applicable).</p>
+    <p>In the below, each message defined in our SRS can be exercised using a button (and input fields when applicable).</p>
     <!-- SERVER TO CLIENT -->
     <h2>Server to Client</h2>
     <h3>Message Available</h3>
