+++ clueless/static/js/main.js
@@ -3,6 +3,33 @@
 WEBSOCKET_URL = 'ws://localhost:8081'; // TODO update to server / something dynamic
 var socket;
 
+var gameState = {
+  players: [],
+  currentPlayer: {},
+  witnessItems: [],
+  turn: false,
+  disqualified: false,
+  microstate: {},
+};
+// TODO change Positions to Position (only one player can move at a time) - Steve
+// TODO improve server response to mock client messages (just change it to a json message) - Steve
+// TODO code to update state when each message received (w/o logic/validation) - Steve, Ken
+// TODO output box for client mock messages - Steve
+// TODO add logging in server so video can show it coming up on the console. - Steve
+// TODO demo: run server with overrides. run tests.
+  // TODO demo:
+  // 1) Video of each message - Rachel
+  // 2) Video of UI so far - Rachel
+  // 3) Voiceover  - Kevin, Phuong
+    // script
+    // record
+
+// timeline:
+  // Steve codes tonight
+  // Rachel does videos tomorrow before class
+  // 9 pm ET: Phuong and Kevin do PowerPoint / voiceover
+  // Video review: async on Tuesday. Turn in by 9 pm
+
 function handleMessage(data) {
   // Handle message received from server
   // TODO handle invalid message that cannot be parsed
@@ -22,10 +49,8 @@ function handleMessage(data) {
   } else if (msg.message == 'Positions') {
     // console.log('Received message Positions');
     $('#msgPositions').html(data);
-
   } else if (msg.message == 'PlayerTurn') {
-    // console.log('Received message PlayerTurn');
-    $('#msgPlayerTurn').html(data);
+    gameState.turn = true;
   } else if (msg.message == 'Suggestion') {
     // console.log('Received message Suggestion');
     $('#msgSuggestion').html(data);
